# ActiveRecord: Improve `find_db_config` performance